external db for ing and recipes is done

+++ recipes-service/src/main/java/domain/model/Ingredient.java
@@ -31,7 +31,7 @@ public class Ingredient implements Serializable {
 	
 	@Column(name="quantite")
 	@NotNull
-	private double quantite;
+	private float quantite;
 	
 	@Column(name="vegetarien")
 	@NotNull
@@ -43,7 +43,7 @@ public class Ingredient implements Serializable {
 	
 	public Ingredient() {}
 
-	public Ingredient(@NotNull Long ingredientId, @NotNull double quantite, @NotNull int vegetarien, @NotNull int vegan) {
+	public Ingredient(@NotNull Long ingredientId, @NotNull float quantite, @NotNull int vegetarien, @NotNull int vegan) {
 		super();
 		this.ingredientId = ingredientId;
 		this.quantite = quantite;
@@ -75,11 +75,11 @@ public class Ingredient implements Serializable {
 		this.ingredientId = ingredientId;
 	}
 
-	public double getQuantite() {
+	public float getQuantite() {
 		return quantite;
 	}
 
-	public void setQuantite(double quantite) {
+	public void setQuantite(float quantite) {
 		this.quantite = quantite;
 	}
